requirements.txt
tensorflow
numpy
pyyaml
termcolor
gym
json_tricks
